data_process/requirements.txt
absl-py==1.4.0
addict==2.4.0
array-record==0.4.0
asttokens==3.0.0
astunparse==1.6.3
attrs==24.3.0
backcall==0.2.0
bcrypt==4.3.0
blessed==1.20.0
blinker==1.8.2
boto3==1.37.18
botocore==1.37.18
cachetools==5.5.2
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cloudpickle==3.1.1
comm==0.2.2
ConfigArgParse==1.7
contourpy==1.1.1
cryptography==44.0.2
cycler==0.12.1
dacite==1.8.1
dash==2.18.2
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
dask==2023.3.1
dataclass_array==1.4.1
decorator==5.1.1
diffusers==0.27.2
dm-tree==0.1.8
einops==0.8.1
einsum==0.3.0
etils==1.3.0
executing==2.1.0
fastjsonschema==2.21.1
filelock==3.16.1
Flask==3.0.3
flatbuffers==25.2.10
fonttools==4.55.3
freetype-py==2.5.1
fsspec==2025.3.0
gast==0.4.0
google-auth==2.16.2
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
googleapis-common-protos==1.63.1
gpustat==1.1.1
grpcio==1.70.0
h5py==3.11.0
huggingface-hub==0.24.7
idna==3.10
imageio==2.35.1
imagesize==1.4.1
immutabledict==2.2.0
importlib_metadata==8.5.0
importlib_resources==6.4.5
ipython==8.12.3
ipywidgets==8.1.5
itsdangerous==2.2.0
jax==0.4.13
jedi==0.19.2
Jinja2==3.1.5
jmespath==1.0.1
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter_core==5.7.2
jupyterlab_widgets==3.0.13
keras==2.11.0
kiwisolver==1.4.7
lark==1.2.2
lazy_loader==0.4
libclang==18.1.1
locket==1.0.0
Markdown==3.7
MarkupSafe==2.1.5
matplotlib==3.6.1
matplotlib-inline==0.1.7
megfile==3.1.7
ml-dtypes==0.2.0
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.1
numpy==1.21.5
numpy-quaternion==2023.0.4
nvidia-ml-py==12.570.86
oauthlib==3.2.2
open3d==0.18.0
opencv-python==4.10.0.84
OpenEXR==1.3.9
opt_einsum==3.4.0
packaging==24.2
pandas==1.5.3
paramiko==3.5.1
parso==0.8.4
partd==1.4.1
pexpect==4.9.0
pickleshare==0.7.5
Pillow==9.2.0
pkgutil_resolve_name==1.3.10
platformdirs==4.3.6
plotly==5.13.1
plyfile==1.0.3
promise==2.3
prompt_toolkit==3.0.48
protobuf==3.19.6
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==10.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pyglet==2.0.20
Pygments==2.18.0
PyNaCl==1.5.0
PyOpenGL==3.1.4
pyparsing==3.1.4
pyquaternion==0.9.9
pyrender==0.1.45
python-dateutil==2.9.0.post0
pytz==2024.2
PyWavelets==1.4.1
PyYAML==6.0.2
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
retrying==1.3.4
rpds-py==0.20.1
rsa==4.9
s3transfer==0.11.4
safetensors==0.5.3
scikit-image==0.20.0
scikit-learn==1.2.2
scipy==1.9.1
six==1.17.0
stack-data==0.6.3
tenacity==9.0.0
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.11.0
tensorflow-addons==0.21.0
tensorflow-datasets==4.9.0
tensorflow-estimator==2.11.0
tensorflow-graphics==2021.12.3
tensorflow-io-gcs-filesystem==0.34.0
tensorflow-metadata==1.13.0
tensorflow-probability==0.19.0
termcolor==2.4.0
terminaltables==3.1.10
threadpoolctl==3.5.0
tifffile==2023.7.10
toml==0.10.2
tomli==2.2.1
toolz==1.0.0
torch==1.8.1
torchvision==0.9.1
tqdm==4.67.1
traitlets==5.14.3
trimesh==4.5.3
typeguard==2.13.3
typing_extensions==4.12.2
tzdata==2024.2
urllib3==1.26.20
visu3d==1.5.1
waymo-open-dataset-tf-2-11-0==1.6.1
wcwidth==0.2.13
Werkzeug==3.0.6
widgetsnbextension==4.0.13
wrapt==1.14.1
yapf==0.43.0
zipp==3.20.2
